[REFACTOR] Ajustado camadas para utilizar corretamente Controller, Service e Raposiries.

+++ src/Service/MainTaskService.cs
@@ -1,16 +1,17 @@
-﻿using Domain.Models;
+﻿using Domain.Mappers;
+using Domain.Models;
+using Domain.Request;
 using Infra.Repositories;
 
 namespace Service;
 
-
 public interface IMainTaskService
 {
-    MainTask Create(MainTask task);
-    void Delete(int taskId);
+    MainTask Create(MainTaskRequest mainTask);
+    void Delete(int mainTaskId);
     List<MainTask>? Get(int userId);
     MainTask? Find(int mainTaskId);
-    MainTask Update(MainTask task, int mainTaskId);
+    MainTask Update(MainTaskUpdate mainTask, int mainTaskId);
 }
 
 public class MainTaskService : IMainTaskService
@@ -22,9 +23,10 @@ public class MainTaskService : IMainTaskService
         _mainTaskRepository = mainTaskRepository;
     }
 
-    public MainTask Create(MainTask task)
+    public MainTask Create(MainTaskRequest mainTaskRequest)
     {
-        return _mainTaskRepository.Create(task);
+        var newMainTask = MainTaskMapper.ToClass(mainTaskRequest);
+        return _mainTaskRepository.Create(newMainTask);
     }
 
     public List<MainTask>? Get(int userId)
@@ -34,15 +36,33 @@ public class MainTaskService : IMainTaskService
 
     public MainTask? Find(int mainTaskId)
     {
-        return _mainTaskRepository.Find(mainTaskId);
+        var mainTask = _mainTaskRepository.Find(mainTaskId);
+
+        if (mainTask is null)
+            throw new Exception("mainTask not found!");
+
+        return mainTask;
     }
 
-    public MainTask Update(MainTask task, int mainTaskId)
+    public MainTask Update(MainTaskUpdate mainTaskUpdate, int mainTaskId)
     {
-        return _mainTaskRepository.Update(task, mainTaskId);
+        var mainTask = _mainTaskRepository.Find(mainTaskId);
+
+        if (mainTask is null)
+            throw new Exception("mainTask not found!");
+
+        mainTask.Description = mainTaskUpdate.Description;
+
+        return _mainTaskRepository.Update(mainTask);
     }
+
     public void Delete(int mainTaskId)
     {
+        var mainTask = _mainTaskRepository.Find(mainTaskId);
+
+        if (mainTask is null)
+            throw new Exception("mainTask not found!");
+
         _mainTaskRepository.Delete(mainTaskId);
     }
